pyproject.toml
[project]
name = "pysurvanalysis"
version = "0.4.0"
description = "Comprehensive survival analysis pipeline for census-based experiments"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lifelines>=0.30.0",
    "matplotlib>=3.10.8",
    "mplcursors>=0.7",
    "numpy>=2.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pyqt6>=6.10",
    "pyqtdarktheme>=2.1.0",
    "pyyaml>=6.0",
    "qtawesome>=1.4.2",
    "scipy>=1.14",
    "statsmodels>=0.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pysurvanalysis"]

[project.scripts]
pysurvanalysis = "main:main"
psurv-hub = "pysurvanalysis.apps.hub:main"
psurv-config = "pysurvanalysis.apps.config_editor:main"
psurv-qc = "pysurvanalysis.apps.qc_viewer:main"
